+++ Backend-Bot-master/app/models/refresh_token.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, TIMESTAMP, func
+from sqlalchemy import BigInteger, String, TIMESTAMP, func
 from sqlalchemy.orm import Mapped, mapped_column
 from app.db import Base
 
@@ -6,8 +6,8 @@ from app.db import Base
 class RefreshToken(Base):
     __tablename__ = 'refresh_tokens'
 
-    id: Mapped[int] = mapped_column(Integer, primary_key=True, autoincrement=True)
-    user_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    id: Mapped[int] = mapped_column(BigInteger, primary_key=True, autoincrement=True)
+    user_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
     token: Mapped[str] = mapped_column(String(255), nullable=False)
     expires_at: Mapped[TIMESTAMP | None] = mapped_column(TIMESTAMP(timezone=True), nullable=True, default=func.now())
     revoked_at: Mapped[TIMESTAMP | None] = mapped_column(TIMESTAMP(timezone=True), nullable=True, default=func.now())
